requirements.txt
jsonschema==3.2.0
aiosqlite==0.13.0
locustio==0.999
